+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.28.0
+
+- **REFACTOR**: Update `ShadCalendar` `yearSelectorMinWidth` from `100` to `64` and `monthSelectorMinWidth` from `120` to `64`.
+- **REFACTOR**: Update `ShadCalendar` position of dropdown, which is now centered in the header.
+- **REFACTOR**: Add `dropdownFormatMonth` and `dropdownFormatYear` to `ShadCalendar`.
+- **FIX**: Center the placeholder in `ShadAvatar`.
+- **DOCS**: Fix missing `MainAxisSize.min` in `ShadCard` notification example.
+- **DOCS**: Adjust sheet docs example, with some spacing between buttons.
+- **FIX**: Fix the hit test behavior of `ShadContextMenuRegion`.
+- **FEAT**: Expose `hitTestBehavior` from `ShadContextMenuRegion`, defaults to `HitTestBehavior.opaque` (thanks to @NonymousMorlock).
+
 ## 0.27.4
 
 - **FIX**: `ShadIconButton` icon size property isn't applied (thanks to @TahaTesser).